Create Graph-Algorithm
Graph algorithms are a fundamental part of computer science. They solve problems related to graph theory in fields such as computer networking, social network analysis, geographical data, and more. In this context, a graph is a set of vertices (or nodes) and edges connecting some pairs of vertices. Graph algorithms aim to solve problems such as pathfinding, connectivity, and flow optimization.