Create Geometric-Algorithms
Geometric algorithms are a class focused on solving geometry-related problems and are a significant area of study in computational geometry. These algorithms deal with objects like points, lines, polygons, and polyhedra. They are often concerned with the relationships among these figures, such as intersection, proximity, connectivity, and optimization of specific parameters.